EZ-3968 - fixing merge conflicts

+++ Integration/Tests/EzBobTest/OpenPlatformTests/Investor/TestInvestorRules.cs
@@ -224,7 +224,7 @@
 
             investorParametersDALMock.Setup(x => x.GetGradeMaxScore(1, (int)Grade.A,1))
                 .Returns((decimal)0.2);
-            investorParametersDALMock.Setup(x => x.GetGradeMonthlyInvestedAmount(1, (int)Grade.A))
+            investorParametersDALMock.Setup(x => x.GetGradeMonthlyInvestedAmount(1, Grade.A))
                 .Returns(100);
 
             investorParametersDALMock.Setup(x => x.GetInvestorMonthlyFundingCapital(1))
